ENH: restructure MonitorGUI location
